Validate that directive arguments are known
