+++ lib/absinthe/pipeline.ex
@@ -138,6 +138,7 @@ defmodule Absinthe.Pipeline do
       {Phase.Schema.Validation.Result, pass: :initial},
       Phase.Schema.FieldImports,
       Phase.Schema.Validation.KnownDirectives,
+      Phase.Document.Validation.KnownArgumentNames,
       {Phase.Schema.Arguments.Parse, options},
       Phase.Schema.Arguments.Data,
       Phase.Schema.Directives,
