+++ test/absinthe/schema/notation/experimental/import_sdl_test.exs
@@ -484,6 +484,41 @@ defmodule Absinthe.Schema.Notation.Experimental.ImportSdlTest do
     end
   end
 
+  test "Validate known directive arguments in SDL schema" do
+    schema = """
+    defmodule SchemaWithDirectivesWithNestedArgs do
+      use Absinthe.Schema
+
+      defmodule Directives do
+        use Absinthe.Schema.Prototype
+
+        directive :some_directive do
+          on [:field_definition]
+        end
+      end
+
+      @prototype_schema Directives
+
+      "
+      type Widget {
+        name: String @some_directive(a: { b: {} })
+      }
+
+      type Query {
+        widgets: [Widget!]
+      }
+      "
+      |> import_sdl
+    end
+    """
+
+    error = ~r/Unknown argument "a" on directive "@some_directive"./
+
+    assert_raise(Absinthe.Schema.Error, error, fn ->
+      Code.eval_string(schema)
+    end)
+  end
+
   def handle_event(event, measurements, metadata, config) do
     send(self(), {event, measurements, metadata, config})
   end
